+++ client/src/components/FootPrint/FootPrint.tsx
@@ -2,9 +2,15 @@ import React from 'react';
 import YellowFootPrint from '../../../public/assets/oneFootPrint.svg';
 import { FootPrintStyle } from './FootPrint.style';
 
-function FootPrint({ width, height }: { width: number; height: number }) {
+function FootPrint({ width, height, fill }: { width?: number; height?: number; fill?: boolean }) {
   return (
-    <FootPrintStyle src={YellowFootPrint} alt="yellowfootPrint" width={width} height={height} />
+    <FootPrintStyle
+      src={YellowFootPrint}
+      alt="yellowfootPrint"
+      fill={fill}
+      width={width}
+      height={height}
+    />
   );
 }
